made subdirectories for migration

+++ database/migrations/Portal/Posts/2019_09_17_193853_create_posts_table.php
@@ -24,7 +24,7 @@ class CreatePostsTable extends Migration
         });
 
         Schema::table('posts', function (Blueprint $table) {
-            $table->integer('user_id')
+            $table->bigInteger('user_id')
                 ->unsigned()
                 ->change();
             $table->foreign('user_id')
